refactor(charts): consolidate LabCharts shared types into shared/types.ts
Move EntityType from LabCharts/types.ts and Granularity from
useGranularity.ts into a single shared/types.ts. Remove re-exports of
EntityType from Top10Race and Top10BillboardRace query files; all
consumers now import directly from shared/types.
