+++ app/src/components/Charts/LabCharts/shared/types.ts
@@ -1,0 +1,3 @@
+export type EntityType = 'artists' | 'tracks' | 'albums'
+
+export type Granularity = 'year' | 'month' | 'week' | 'day'
