--- app/src/components/Charts/LabCharts/types.ts
@@ -1,1 +1,0 @@
-export type EntityType = 'artists' | 'tracks' | 'albums'
